requirements.txt
discord.py>=2.3
aiohttp>=3.8
Flask>=2.0
requests>=2.28
python-dotenv>=1.0